// src/components/Form/Form.types.ts
export interface IFormInput {
	elternVorname: string;
	elternNachname: string;
	elternEmail: string;
	elternMobilnummer: string;
	kindVorname: string;
	kindNachname: string;
	kindGeburtstag: string;
	strasse: string;
	hausnummer: string;
	stadt: string;
	plz: string;
	message: string;
	start: string;
	ende: string;
}